requirements.txt
Flask==1.0.2
Flask-Login==0.4.1
Flask-MySQL==1.4.0
Jinja2==2.1
MarkupSafe==1
Pillow==5.1.0
PyMySQL==0.8.1
WTForms==2.1
Werkzeug==0.14.1
aniso8601==3.0.0
click==6.7
flask-restplus==0.11.0
itsdangerous==0.24
jsonschema==2.6.0
pip==10.0.1
pytz==2018.4
setuptools==39.2.0
six==1.11.0
